Return string instead of Response
Returning a Response object leads to having a response nested in a response.
Django resolves it by iterating over the nested response and extracting the content.
But the problem is that Response has a 'close()' method, which Django calls.
This closes the database connnection while the request is still ongoing.
This leads to errors when opening the API response in a testcase.